requirements.txt
onnxruntime
opencv-python
firebase-admin
numpy
Pillow
torch
torchvision
picamera2